% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{porpoiseProb.xgb}
\alias{porpoiseProb.xgb}
\title{Filter a PAL dataset}
\usage{
porpoiseProb.xgb(test)
}
\arguments{
\item{newdata}{A data frame containing a PAL dataset (see vignette)}
}
\value{
A numeric vector of predicted porpoise probabilities
}
\description{
Filter a PAL dataset
}
